Move timeline and events into own folder and update imports
